Ability Rotation/Moving Platforms
Ability needs to be hooked up to a renderable object currenlty, that is if you want to implement it. I still need to fix levers tommorow if we want those in levels too.

I also added a player light and scene lighting stuff, We can mess with that more on level design! It may be cool to have facility lights change color rather than the whole screen tent